Break Firebase cycle
