+++ config/firebaseClient.js
@@ -1,0 +1,16 @@
+import { initializeApp, getApp } from 'firebase/app';
+import { getFirestore } from 'firebase/firestore';
+import { firebaseConfig } from './firebase';
+
+let app;
+try {
+  app = initializeApp(firebaseConfig);
+} catch (error) {
+  if (error.code !== 'app/duplicate-app') {
+    console.error('Firebase initialization error:', error);
+  }
+  app = getApp();
+}
+
+export const db = getFirestore(app);
+export { app };
